Char#digit?: Use nt field instead of relying on Nd
We'd previously defined #digit? as returning true if the character had the
general category of Nd. The UCD XML dump has an nt field (Numeric_Type), a
possible value of which is 'Digit'). It's more precise to use this property
rather than attempt to derive it ourselves. Further, Unicode recommend
against deriving such properties algorithmically.